+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[UPDATE]** Add ariaLabel attribute to `ItemInfo` and `Item`.
+- **[FIX]** Improve a11y on `Stepper` for keyboard navigation and screen readers
 - [..]
 
 # v22.2.0 (06/03/2020)